Make methods exposed by Container match those exposed by Window.
Conceptually speaking, a Container can be considered to be a parallel Window.  That is, it permits many of the same operations: adding an existing Window, defining a new Window etc. As such the methods exposed by container should match those exposed by Window as closely as possible.  Accordingly, this change adds new Container#over, Container#use (replacing Container#also_use) and Container#repeat methods.  These methods match the equivalent methods on the Window class with the exception that they permit an `after` parameter.  This can be used to stagger any parallel timelines added to the Container (i.e. so that they don't start at the same time.)