Use the Container API directly when configuring parallel moments from a Window
In the previous commit (4fc4f6417c2b21733b3c53fd3b3af3149e093576) the Container API was adjusted so that it more closely mirrors that of the Window.  This change updates the Window#simultaneously method to accept a block whose contents will be evaluated within the context of a newly created Container object before being appended to the Window.  This makes for a far simpler, more consisted API.