Remove all references to `cram`ming
The `cram` parameter was added to allow for simple scaling of a timeline by having every Moment configured be scheduled multiple times.  This, however, intriduces a lot of irregularities: for example, consider the following window:

```ruby
Dodo.over 2.weeks do
  please do
    some_var = []
  end

  please do
    some_var << Time.now
   end
 end
```

This will create two Array objects and perform two append actions.  However, _both_ append actions will be sent to the same Array object (the one most recently created).  This is not really desirable and renders this particular way of scaling unusable for many applications.

Furthermore, the code controlling this scaling was particularly confusing and shitty in general.

This commit nukes the lot of it.

A better solution will be to simply parallelise the root level window when we want to scale.  This will allow integral multitples of the total number of moments in a given Window to be evaluated and should avoid the issue described above.