Comment out flaky tests
Since we removed the `cram` parameter in 9274972ee1657b62d226ea57af6f3c5513f89376 there is no way to increase the total number of `Moment`s included in the test cases within the `with rand achieving its upper bound` context.  This means that it is now probable that at least one of these tests will fail if run enough times.  As such, these tests should be commented until a new equivalent to `cram` is established.