Refactor the progress classes to make them less confusing
The `ProgressCounter` module makes use of `current=` and `+` methods which behave in a slightly surprising way.  Rename and combine these methods in order to make them more descriptive.  Also include a ProgressCounter::Unknown 'null' object to encapsulate logic pertaining to a missing progress total (I expect this may neem extending soon).

Finally, provide a `factory` method on the `ProgressCounter` singleton class that will be used to create a callable that can be used to create a `ProgressCounter` object at a later point.  This way we can keep progress totals immutable (we would never want this to change) and avoid exposing the total= method.