Add a schedule_with method to Happening
In order to DRY up the Happening subclasses a `schedule_with` method has been added that can be used to specify the associated scheduler class.  Since the system as a whole depends on these classes having the same interface this abstraction is well defined.  Finally, we could make the Happening subclasses _even_ simpler by using the convention that each associated `Scheduler` class is called `<Happening subclass name>Scheduler` but, personally I dislike the 'Convention Over Configuration' approach and prefer that the choice of scheduler class be explicit.