Ensure uniformity of Happening classes
It makes sense to think of a Happening as always having an ordered list
of child happenings.  Doing so allows us to treat Windows, Containers and
Moments more uniformly.  This commit moves the creation of an empty
happening array into the Happening base class initializer and exposes an
each method on every Happening that will iterate through these child
happenings.  This will be the only way to access these happenings