Expose addition and scalar multiplication and paralellisation operators on the Window class
each of these operators returns a new `Window` object. the addition (`+`) operator creates a new `Window` whose child timelines are the concatenation of the two original windows and whose duration is the sum of the durations of the two original windows.  The scalar multiplication operator performs repeated addition.  The "parallelisation operator" (tentatively (`**`) at the moment) creates a new window whose duration is that of the original window and that consists of a single container, parallelising the original window `n` times (where `n` is the integral argument passed to the operator).

The idea is that these operators can be used to create scalable timelines that can be used to generate datasets of varying size, as determined at runtime.

For example, consider a situation in which we wish to generate data corresponding to an organisation and then mock sales data belonging that organisation.  Suppose also that we wish to be able to generate datasets of different sizes.

Regardless of the size of the dataset we would only ever want one organisation with the size of the datasets then differing in the quantities of the sales data only.

In the code below, let `organisation_window` be a `Window` which could be used to create data pertaining to the organisation data itself while `sales_window` be a window to create individual sales item data.  Finally let `factor` denote some integral factor which we may want to scale by.  This could be provided as an argument in a rake task.

```ruby
scaled_window = organisation_window + (factor * sales_window)
```