Restore tests of scaled windows.
This commit restores the test removed in 9274972ee1657b62d226ea57af6f3c5513f89376
scaling using the new `**` operator added in cff8c6d6df2d306cdfb9ec50ae76f604a750b0cf

Certain time based test cases are expected to fail.  The reason why is
best illustrated with a direct reference to the following test case.
By paralellising the `window` using the ** operator we are _effectively_
scheduling the window twice and sorting by offset.  As such, the time
at which the first author is created in one of the parallelised 'versions'
of the window may be very different from the time the first author is
created in the other 'version' despite the fact that in both cases the
authors required were created within the specified time frame.

Skipping the tests until I can think of how to improve them.