Ensure Scope#respond_to? method behaves as expected
The `Scope#response_to_missing?` method was not being covered in tests.  Furthermore, attempting to read send a message to a scope that was not a valid instance variable when prepended with a `@` (e.g. `:matches?`) resulted in `NameError`s.  This commit ensures attempts to send such messages are handled correctly by ensuring that their handling is deferred to the superclass of `Scope`